Num inference

+++ packages/emitter/testcases/arrays/basic/ArrayLiteral.cs
@@ -1,10 +1,14 @@
+// Generated from: ArrayLiteral.ts
+// Generated at: 2025-12-13T16:22:31.272Z
+// WARNING: Do not modify this file manually
+
 namespace TestCases.arrays.basic
 {
         public static class ArrayLiteral
         {
             public static global::System.Collections.Generic.List<double> createArray()
                 {
-                var arr = new global::System.Collections.Generic.List<double> { 1.0, 2.0, 3.0 };
+                var arr = new global::System.Collections.Generic.List<int> { 1, 2, 3 };
                 return arr;
                 }
         }
