+++ packages/emitter/testcases/arrays/multidimensional/MultiDimensional.cs
@@ -1,3 +1,7 @@
+// Generated from: MultiDimensional.ts
+// Generated at: 2025-12-13T16:22:31.325Z
+// WARNING: Do not modify this file manually
+
 namespace TestCases.arrays.multidimensional
 {
         public static class MultiDimensional
@@ -9,7 +13,7 @@ namespace TestCases.arrays.multidimensional
 
             public static global::System.Collections.Generic.List<global::System.Collections.Generic.List<double>> createMatrix()
                 {
-                return new global::System.Collections.Generic.List<global::System.Collections.Generic.List<double>> { new global::System.Collections.Generic.List<double> { 1.0, 2.0 }, new global::System.Collections.Generic.List<double> { 3.0, 4.0 } };
+                return new global::System.Collections.Generic.List<global::System.Collections.Generic.List<double>> { new global::System.Collections.Generic.List<int> { 1, 2 }, new global::System.Collections.Generic.List<int> { 3, 4 } };
                 }
         }
 }
